DPL: move topological sort in a separate file
Simplifies testing.
